Minor refactoring of RISC-V side HDL, fixes some bugs found in synthesis.
Namely, a problem in the previous commit was that HREADY was constantly asserted and deasserted, which trashed the bus and did not allow GRMON to function properly.
Also renames some files to shorten them, removes some unnecessary files (.swp) and toplevel modules. Instead of hierarching the entire ahb-unit, it is now flattened
(pico-adapter-freeahb).

We also get started with some system software by utilizing existing sources provided by peers, but it has not been tested on chip yet.