Lends some LEDs and a button from GRLIB's GPIO-unit to ease debugging.
Namely, we add a button for resetting the pico parts of the design (but this can only be done when it is not granted the bus,
as the protocol doesnt really support just quitting without any notice).

DS15 (LED) is mapped to this new button. DS16 is mapped to the Pico's FreeAHB's HBUSREQ. DS17 is mapped to the incoming GRANT.

Not entirely sure what did it, but HBUSREQ is now not raised constantly anymore (we first tried inserting wait states, but after removing most - if not all- of them,
there is still no constant HBUSREQ), and we do therefore not lock others (including the JTAG) out of the bus.

Should do a git diff at some point...

We need to continue with systemwide simulation in order to ensure correct operation. We also need to simulate a run with test software,
as we firmly believe that the endianness is out of whack (this AMBA bus targets LEON3's big endian, while RISC-V is little endian).