Fixes bugs in freeahb_adapter and its TB.
Namely, there was an error in the write FSM, in which disabled bytes caused the adapter to reuse earlier write signals, writing them double.
Not necesserily a severe bug, but it causes the FreeAHB master to cause unneccessary strain to the bus. We have now fixed this behavior.

Another error was in the FSM as a whole. It takes at least one cycle from raising mem_ready, to mem_valid being lowered. This caused
in our TB our read FSM to start over once. Ideally, we would like to raise freeahb_valid for one cycle, then start preparing the next transfer,
but in its current implementation, FreeAHB requires that the UI remains unchanged when freeahb_next is not raised. Therefore, we utilize a
transfer_done register to halt the adapter from making further transfers, until mem_valid is lowered, which clears this register.

This commit also adds a simple script which runs the testbenches.

The TB with the adapter and FreeAHB combined as the DUV is broken, this will be fixed in an upcomming commit.