Fixes write ctr bug and testbench bug.
Specifically, there was a default else for writes which should not have been a default, causing the counter to tick on every clock even though it should not.
Also makes freeahb_next a bit element instead of a logic element, since elsewise it breaks some of the checks we do in our FSM, which assume binary logic, not tristate.