Fixes errors in the adapter.
Namely, we now latch read data, so that we only save the value which is given on HREADY, and not at any other time.
We also set freeahb_min_len to 0, as elsewise we get a INCR16 read for a single read, which causes faulty reads, as we expect our reads
to be a single 32-bit INCR with one beat read (single transfer). We also hardcode our assumption that 1 cycle of mem_ready should result in mem_valid being lowered, by reflecting the expected state.