Changes Pico core config in GRLIB design, and fixes write order in memory adapter.
The pico core config is changed to more closely resemble the configurations we have in the TB we derived from the original PicoRV software testbench.
The memory adapter did writes incorrectly for big endian. The addresses that are written to are the reverse of that of a little endian target.