Patches adapter to respect the FreeAHB's o_next signal properly.
Namely, we did a wrongful assumption up until now that i_dav
(which we renamed freeahb_valid), is to indicate to the FreeAHB that the UI is valid.

The original name, and a quick look at the source code, proves that this is not
the case. FreeAHB UI is assumed to be valid if i_read or i_write and i_dav/freeahb_valid
is asserted, and MUST be changed on freeahb_next/o_next if one does not wish to
duplicate writes and reads.

A read is successful on freeahb_ready/o_dav, a write is successful on freeahb_next.

Since our adapter design solidly depends on that there is only one instance of every
read and write, as it never checks if it is in sync with the processor and/or FreeAHB,
it proved to be a crucial misassertion in realistic test environments.

The design now passes the fulltest TB again, and hopefully it will now work in a
GRLIB environment as well.