Fixes pipelining behavior of adapter, and more accurate full testbench.
Finally found the two biggest dum-dums which I've introduced.

First, the reason why there was one observed faulty behavior in the GRLIB testbench in Active HDL, and not one in the full_testbench running on Icarus Verilog,
is due to an errata in the TB. We - by mistake - used mem_wdata for writing to memory, instead of the actual o_hwdata, the output of freeAHB to the AHB bus, which therefore never tested
the interaction between the adapter and FreeAHB, and therefore we got this error.

The second, which led to strange or missing writes (on accurate testbenches)  was the misunderstanding of the function of the FreeAHB UI.
The UI functions just as one should expect from the AHB spec: initialize the AHB pipeline with an initial address phase in the UI, then follow on with its data when freeahb_next.
For some reason we had concluded that most of the pipelining was abstracted away via the UI, but with some sceptic thought and a quick dive into the FreeAHB source, you'd
see that the pipelined behavior is indeed expected on the UI.

Tests successful with a simulated memory in iverilog, now to test on the GRLIB simulator on ActiveHDL.