Commits the small modifications done to the SW-chain for working prototype.
It is working in a very loose sense of the word, as there is still problems writing to the UART
since it isn't really connected to anything on the board. Currently working on a JTAG application
for polling the UART in debug mode.

The _write newlib call now writes to a single byte in memory. This will be changed to the UART soon.
Other updates is to make srecs with more zeropadding, as this is assumed in the test program. The
program start is elsewhere, the program start, and the SP has also been relocated slightly.