First prototype of uartMonitor.
It sets APBUART in debug mode. Then, it constantly polls the status register.
When the APBUART status register has the TE (Transmitter FIFO empty) bit set,
it will read out the value of the transmitter FIFO via the FIFO debug register.

The printout is a bit verbose at the moment, but will be cleaned up after doing
some testing with PicoRV prints. The issue at the moment is that printf waits
until a newline to print to our console. During debug we wont neccesarily see a
newline character in our UART write stream.