Modifies riscv.ld to also link start.S
The interesting parts of start.S is now incorporated into riscv.ld.
The stack pointer base address is given its own symbol stackPointerAddr,
so that one doesn't need to manually edit the start.S file to set the correct SP.

Entrypoint of riscvld is set to the _pvstart symbol, which is the start.S assembly.

Makefile is edited so that start.S is compiled into object code, but not linked before
assembling firmware.srec.