Readies the toplevel design for interrupts.
Namely, we share button1 (enable) with both the Pico for enabling, and as a GPIO button.
The board only has two available user buttons, hence the sharing.

In addition, GRGPIO is reconfigured for this new input width, and we also set a single interrupt line,
and we mask so that only button1 is the button in use. We remove the configuration variables for GRGPIO,
since we do not have the time to make the design configurable, and therefore the design becomes more
readable by having values directly in the leon3mp.vhd