First working version of the design WITH interrupts :D
Much obliged due to the already existing works from the original Pico repository,
and the GRLIB documentation.

The start.S does not wait for a start interrupt, and enables all interrupts straight away.
In firmware.cc, the UART is immediately put in debug mode (just in case, but shouldnt really be necessary at this point),
and then a welcome message is printed. Then, GRGPIO is setup so that the enable button can trigger an interrupt.

After the peripheral setup, we can interrupt the execution at any time.
The only downside is if one interrupts during a print, as the original write buffer will be overwritten by the handler. :(