Adds UART enquiry support.
Both firmware and UartMonitor's UART configuration is updated to enable
the APBUART's control register so that receiver interrupts are enabled.

The syscalls' _read stub is updated to request and fetch one byte
from the UART. It does so by writing a ASCII ENQ byte to signal an enquiry,
then it waits for any interrupt before it fetches a character from
the UART receiver FIFO. This could be improved upon by checking which
interrupts are actually handled. Or one could check the receiver count
and see if there actually is a byte there.

The firmware.cc has been updated to print more informative startup messages.
It also calls getchar in order to get a character from the user when needed.

Next, we will fix timer interrupts. Then append a small game to the end of firmware.cc,
then we will consider our prototype done. The thesis report doesn't write itself.