Adds timer interrupt support, slims IRQ handler, makes firmware prints cleaner.
Final version of the proof of concept. GRLIB GPTIMER is now configured to have only one timer,
which triggers interrupt line 8 when underflowing. The firmware sets the prescaler and timer value register
so that the clock should be interrupting every second. We avoid having prints in the timer interrupt handling,
as the current UART print flow via AHBJTAG adds significant latency. We also clean up the prints that the user
sees in the UartMonitor in general, so that the demo program overall is cleaner.