Gets testbenches working again, by separating the SW built for implementation, and for simulation.
In short, we do not add any support for interrupt handling testing or system configuration,
as the firmware for the implementation. This is purely due to time constraints.
So we stick to the pre-IRQ SW for the testbenches so we don't have to extend them. This would
require us to start adding system stubs for the rest of the system and GRLIB peripherals, and
we would in that case like to do it properly or not at all.

picorv/software therefore gets a new folder, sim, where all the simulation software resides.
Affected files and toolchains are updated to reflect this change. See the git diff.